@@ -1,74 +1,56 @@
 {
-  "name": "star-angular",
-  "version": "2.0.0",
-  "license": "MIT",
+  "name": "star-admin-angular",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "webpack-dev-server --port=4200",
-    "build": "webpack",
-    "test": "karma start ./karma.conf.js",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
     "lint": "ng lint",
-    "e2e": "protractor ./protractor.conf.js",
-    "pree2e": "webdriver-manager update --standalone false --gecko false --quiet"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/router": "^5.2.11",
-    "@ng-bootstrap/ng-bootstrap": "1.0.1",
-    "bootstrap": "^4.1.1",
-    "compass-mixins": "0.12.10",
-    "core-js": "^2.5.7",
-    "font-awesome": "4.7.0",
-    "mdi": "2.1.99",
-    "rxjs": "^5.5.11",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^3.3.0",
+    "bootstrap": "^4.1.3",
+    "compass-mixins": "^0.12.10",
+    "core-js": "^2.5.4",
+    "font-awesome": "^4.7.0",
+    "mdi": "^2.2.43",
+    "rxjs": "~6.2.0",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/core": "0.3.2",
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/language-service": "^5.2.11",
-    "@ngtools/webpack": "1.10.2",
-    "@types/jasmine": "^2.8.7",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.111",
-    "autoprefixer": "^7.2.3",
-    "circular-dependency-plugin": "^4.2.1",
-    "codelyzer": "^4.3.0",
-    "copy-webpack-plugin": "~4.4.1",
-    "file-loader": "^1.1.5",
-    "html-webpack-plugin": "^2.29.0",
-    "istanbul-instrumenter-loader": "^3.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "~6.2.5",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.3.0",
+    "file-loader": "^2.0.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.2",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "less-loader": "^4.0.5",
-    "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.1.5",
-    "postcss-url": "^7.3.2",
-    "protractor": "~5.1.2",
-    "raw-loader": "^0.5.1",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.19.1",
-    "stylus-loader": "^3.0.1",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "url-loader": "^0.6.2",
-    "webpack": "~3.11.0",
-    "webpack-dev-server": "~2.11.0"
+    "protractor": "~5.4.0",
+    "sass-loader": "^7.1.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2"
   }
 }
